Moved CommonButtonBase to the new Base folder and renamed it to MudBaseButtonComponent. Changed all components that was using it.
