+++ src/MudBlazor/Components/Button/Fab.razor.cs
@@ -7,7 +7,7 @@ using MudBlazor.Extensions;
 
 namespace MudBlazor
 {
-    public class ComponentBaseFab : CommonButtonBase
+    public class ComponentBaseFab : MudBaseButtonComponent
     {
         protected string Classname =>
         new CssBuilder("mud-fab-root mud-fab")
