+++ src/MudBlazor/Components/Button/IconButton.razor.cs
@@ -7,7 +7,7 @@ using MudBlazor.Extensions;
 
 namespace MudBlazor
 {
-    public class ComponentBaseIconButton : CommonButtonBase
+    public class ComponentBaseIconButton : MudBaseButtonComponent
     {
         protected string Classname =>
         new CssBuilder("mud-button-root mud-icon-button")
